+++ Runtime/Views/VoiceOverPresenter.cs
@@ -140,7 +140,7 @@ namespace Yarn.Unity
             {
                 await YarnTask.Delay(
                     TimeSpan.FromSeconds(waitTimeBeforeLineStart),
-                    lineCancellationToken.NextLineToken).SuppressCancellationThrow();
+                    lineCancellationToken.NextContentToken).SuppressCancellationThrow();
             }
 
             // Start playing the audio.
@@ -148,7 +148,7 @@ namespace Yarn.Unity
 
             // Playback may not begin immediately, so wait until it does (or if
             // the line is interrupted.)
-            await YarnTask.WaitUntil(() => audioSource.isPlaying, lineCancellationToken.NextLineToken).SuppressCancellationThrow();
+            await YarnTask.WaitUntil(() => audioSource.isPlaying, lineCancellationToken.NextContentToken).SuppressCancellationThrow();
 
             if (!DialogueRunner.IsInPlaymode)
             {
@@ -157,7 +157,7 @@ namespace Yarn.Unity
 
             // Now wait until either the audio source finishes playing, or the
             // line is interrupted.
-            await YarnTask.WaitUntil(() => !audioSource.isPlaying, lineCancellationToken.NextLineToken).SuppressCancellationThrow();
+            await YarnTask.WaitUntil(() => !audioSource.isPlaying, lineCancellationToken.NextContentToken).SuppressCancellationThrow();
 
             if (!DialogueRunner.IsInPlaymode)
             {
@@ -167,7 +167,7 @@ namespace Yarn.Unity
             // If the line was interrupted while we were playing, we need to
             // wrap up the playback as quickly as we can. We do this here with a
             // fade-out to zero over fadeOutTimeOnLineFinish seconds.
-            if (audioSource.isPlaying && lineCancellationToken.IsNextLineRequested)
+            if (audioSource.isPlaying && lineCancellationToken.IsNextContentRequested)
             {
                 // Fade out voice over clip
                 float lerpPosition = 0f;
@@ -196,11 +196,11 @@ namespace Yarn.Unity
             // because the user has already indicated that they're fine with
             // things moving faster than sounds normal.)
 
-            if (!lineCancellationToken.IsNextLineRequested && waitTimeAfterLineComplete > 0)
+            if (!lineCancellationToken.IsNextContentRequested && waitTimeAfterLineComplete > 0)
             {
                 await YarnTask.Delay(
                     TimeSpan.FromSeconds(waitTimeAfterLineComplete),
-                    lineCancellationToken.NextLineToken
+                    lineCancellationToken.NextContentToken
                 ).SuppressCancellationThrow();
             }
 
@@ -229,11 +229,6 @@ namespace Yarn.Unity
             return YarnTask.CompletedTask;
         }
 
-        /// <inheritdoc/>
-        public override YarnTask<DialogueOption?> RunOptionsAsync(DialogueOption[] dialogueOptions, CancellationToken cancellationToken)
-        {
-            return DialogueRunner.NoOptionSelected;
-        }
         /// <inheritdoc/>
         public override YarnTask OnDialogueStartedAsync()
         {